created a utility file and move the previous loss utilities code from variational.py to it. converted the methods to static methods.
