+++ src/qibo/models/utils.py
