+++ src/qibo/models/variational.py
@@ -3,7 +3,7 @@ import numpy as np
 
 from qibo.config import raise_error
 from qibo.models.evolution import StateEvolution
-from qibo.models.utilities import Loss_Utils
+from qibo.models.utils import Loss_Utils
 
 
 class VQE(object):
